Add functionality to edit values of date and time parameters
Add the functionality to edit the date and time parameters when button set is
pressed. The SW only starts monitoring this button after we enter in edition
mode.

The value is only updated on the LCD, for the moment the SW doesn't update the
date and time in the RTC module. This is going to be implemented in the next
commits.