Fix: String index for values of the days of the week
There was a mismatch between the range for the days of the week and position
of the strings in the array. The value of 1 corresponds to the first day of the
week, which in the program is Monday.

Also the minimum and maximum values for the days of the week in the function
increment_param was updated to match the valid range accepted by the RTC module.